+++ pkg/backend/httpstate/client/client.go
@@ -380,7 +380,7 @@ func (pc *Client) GetStackUpdates(ctx context.Context, stack StackIdentifier) ([
 // ExportStackDeployment exports the indicated stack's deployment as a raw JSON message.
 // If version is nil, will export the latest version of the stack.
 func (pc *Client) ExportStackDeployment(
-	ctx context.Context, stack StackIdentifier, version *int) (apitype.UntypedDeployment, error) {
+	ctx context.Context, stack StackIdentifier, version *int, showSecrets bool) (apitype.UntypedDeployment, error) {
 
 	path := getStackPath(stack, "export")