+++ pkg/resource/stack/deployment_test.go
@@ -86,7 +86,7 @@ func TestDeploymentSerialization(t *testing.T) {
 		nil,
 	)
 
-	dep, err := SerializeResource(res, config.NopEncrypter)
+	dep, err := SerializeResource(res, config.NopEncrypter, false /* showSecrets */)
 	assert.NoError(t, err)
 
 	// assert some things about the deployment record:
@@ -182,7 +182,7 @@ func TestUnsupportedSecret(t *testing.T) {
 	rawProp := map[string]interface{}{
 		resource.SigKey: resource.SecretSig,
 	}
-	_, err := DeserializePropertyValue(rawProp, config.NewPanicCrypter())
+	_, err := DeserializePropertyValue(rawProp, config.NewPanicCrypter(), config.NewPanicCrypter())
 	assert.Error(t, err)
 }
 
@@ -190,7 +190,7 @@ func TestUnknownSig(t *testing.T) {
 	rawProp := map[string]interface{}{
 		resource.SigKey: "foobar",
 	}
-	_, err := DeserializePropertyValue(rawProp, config.NewPanicCrypter())
+	_, err := DeserializePropertyValue(rawProp, config.NewPanicCrypter(), config.NewPanicCrypter())
 	assert.Error(t, err)
 }
 
@@ -288,7 +288,7 @@ func TestCustomSerialization(t *testing.T) {
 	// Using stack.SerializeProperties will get the correct behavior and should be used
 	// whenever persisting resources into some durable form.
 	t.Run("SerializeProperties", func(t *testing.T) {
-		serializedPropMap, err := SerializeProperties(propMap, config.BlindingCrypter)
+		serializedPropMap, err := SerializeProperties(propMap, config.BlindingCrypter, false /* showSecrets */)
 		assert.NoError(t, err)
 
 		// Now JSON encode the results?